+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Current (unreleased)
 
-- Improved "initial data import" filter/selector, according to the suggestions in https://github.com/riparias/gbif-alert/issues/251.
+- Improved "initial data import" filter/selector, according to the suggestions in https://github.com/riparias/gbif-alert/issues/251
+- Internal improvement to improve the tool re-usability (https://github.com/riparias/gbif-alert/issues/250)
 
 # v1.2.1 (2023-08-21)