+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Current (unreleased)
 
 - Users can now easily change their password
+- The vernacular name is now shown in the observations table
 - Improved "initial data import" filter/selector, according to the suggestions in https://github.com/riparias/gbif-alert/issues/251
 - Internal improvement to improve the tool re-usability (https://github.com/riparias/gbif-alert/issues/250)
 - Update dependencies